+++ shardingsphere-features/shardingsphere-authority/shardingsphere-authority-common/src/main/java/org/apache/shardingsphere/authority/engine/AuthenticationContext.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.authority.auth;
+package org.apache.shardingsphere.authority.engine;
 
 import lombok.AccessLevel;
 import lombok.Getter;
